+++ constants.js
@@ -19,7 +19,7 @@ const INTERFACE_IDS = {
 	ERC725Y: '0x714df77c',
 	LSP0ERC725Account: '0x9a3bfe88',
 	LSP1UniversalReceiver: '0x6bb56a14',
-	LSP1UniversalReceiverDelegate: '0xc2d7bcc1',
+	LSP1UniversalReceiverDelegate: '0xa245bbda',
 	LSP6KeyManager: '0xc403d48f',
 	LSP7DigitalAsset: '0xe33f65c3',
 	LSP8IdentifiableDigitalAsset: '0x49399145',
@@ -86,19 +86,28 @@ const ERC725YKeys = {
 		// keccak256('LSP4Metadata')
 		LSP4Metadata: '0x9afb95cacc9f95858ec44aa8c3b685511002e30ae54415823f406128b85b238e',
 		// keccak256('"LSP4Creators[]')
-		'LSP4Creators[]': '0x114bd03b3a46d48759680d81ebb2b414fda7d030a7105a851867accf1c2352e7',
+		'LSP4Creators[]': {
+			// use this key to get the number of elements in the array
+			length: '0x114bd03b3a46d48759680d81ebb2b414fda7d030a7105a851867accf1c2352e7',
+			// use this key + bytes16(index) to access an index in the array
+			index: '0x114bd03b3a46d48759680d81ebb2b414',
+		},
 	},
 	LSP5: {
 		// LSP5ReceivedAssetsMap:<address>
 		LSP5ReceivedAssetsMap: '0x812c4334633eb816c80d0000',
 		// keccak256('LSP5ReceivedAssets[]')
-		'LSP5ReceivedAssets[]':
-			'0x6460ee3c0aac563ccbf76d6e1d07bada78e3a9514e6382b736ed3f478ab7b90b',
+		'LSP5ReceivedAssets[]': {
+			length: '0x6460ee3c0aac563ccbf76d6e1d07bada78e3a9514e6382b736ed3f478ab7b90b',
+			index: '0x6460ee3c0aac563ccbf76d6e1d07bada',
+		},
 	},
 	LSP6: {
 		// keccak256('AddressPermissions[]')
-		'AddressPermissions[]':
-			'0xdf30dba06db6a30e65354d9a64c609861f089545ca58c6b4dbe31a5f338cb0e3',
+		'AddressPermissions[]': {
+			length: '0xdf30dba06db6a30e65354d9a64c609861f089545ca58c6b4dbe31a5f338cb0e3',
+			index: '0xdf30dba06db6a30e65354d9a64c60986',
+		},
 		// AddressPermissions:Permissions:<address>
 		'AddressPermissions:Permissions': '0x4b80742de2bf82acb3630000',
 		// AddressPermissions:AllowedAddresses:<address>
@@ -114,13 +123,19 @@ const ERC725YKeys = {
 		// keccak256('LSP10VaultsMap')
 		LSP10VaultsMap: '0x192448c3c0f88c7f238c0000',
 		// keccak256('LSP10Vaults[]')
-		'LSP10Vaults[]': '0x55482936e01da86729a45d2b87a6b1d3bc582bea0ec00e38bdb340e3af6f9f06',
+		'LSP10Vaults[]': {
+			length: '0x55482936e01da86729a45d2b87a6b1d3bc582bea0ec00e38bdb340e3af6f9f06',
+			index: '0x55482936e01da86729a45d2b87a6b1d3',
+		},
 	},
 	LSP12: {
 		// LSP12IssuedAssetsMap:<address>
 		LSP12IssuedAssetsMap: '0x74ac2555c10b9349e78f0000',
 		// keccak256('LSP12IssuedAssets[]')
-		'LSP12IssuedAssets[]': '0x7c8c3416d6cda87cd42c71ea1843df28ac4850354f988d55ee2eaa47b6dc05cd',
+		'LSP12IssuedAssets[]': {
+			length: '0x7c8c3416d6cda87cd42c71ea1843df28ac4850354f988d55ee2eaa47b6dc05cd',
+			index: '0x7c8c3416d6cda87cd42c71ea1843df28',
+		},
 	},
 };
 
@@ -141,7 +156,7 @@ const BasicUPSetup_Schema = [
 	},
 	{
 		name: 'LSP12IssuedAssets[]',
-		key: ERC725YKeys.LSP12['LSP12IssuedAssets[]'],
+		key: ERC725YKeys.LSP12['LSP12IssuedAssets[]'].length,
 		keyType: 'Array',
 		valueContent: 'Number',
 		valueType: 'uint256',
@@ -172,6 +187,136 @@ const PERMISSIONS = {
     SUPER_STATICCALL:     "0x0000000000000000000000000000000000000000000000000000000000002000", // 0010 .... .... ....
     SUPER_DELEGATECALL:   "0x0000000000000000000000000000000000000000000000000000000000004000", // 0100 .... .... ....
 }
+
+const LSP1_TYPE_IDS = {
+	// keccak256('LSP7TokensSender')
+	LSP7_TOKENSENDER: "0x40b8bec57d7b5ff0dbd9e9acd0a47dfeb0101e1a203766f5ccab00445fbf39e9",
+	// keccak256('LSP7TokensRecipient')
+	LSP7_TOKENRECIPIENT: "0xdbe2c314e1aee2970c72666f2ebe8933a8575263ea71e5ff6a9178e95d47a26f",
+	// keccak256('LSP8TokensSender')
+	LSP8_TOKENSENDER: "0x3724c94f0815e936299cca424da4140752198e0beb7931a6e0925d11bc97544c",
+	// keccak256('LSP8TokensRecipient')
+	LSP8_TOKENRECIPIENT: "0xc7a120a42b6057a0cbed111fbbfbd52fcd96748c04394f77fc2c3adbe0391e01",
+	// keccak256("LSP9VaultSender")
+	LSP9_VAULTSENDER: "0x3ca9f769340018257ac15b3a00e502e8fb730d66086f774210f84d0205af31e7",
+	// keccak256("LSP9VaultRecipient")
+	LSP9_VAULTRECIPIENT: "0x09aaf55960715d8d86b57af40be36b0bfd469c9a3643445d8c65d39e27b4c56f"
+}
+
+// ----------
+
+const Errors = {
+	LSP6: {
+		'0xf292052a': {
+			error: 'NoPermissionsSet(address)',
+			message: 'LSP6: No permissions are set for this address.',
+		},
+		'0x3bdad6e6': {
+			error: 'NotAuthorised(address,string)',
+			message: 'LSP6: Not authorized (missing permission).',
+		},
+		'0x2879b128': {
+			error: 'NotAllowedAddress(address,address)',
+			message: 'LSP6: not allowed to interact with address.',
+		},
+		'0x84135110': {
+			error: 'NotAllowedFunction(address,bytes4)',
+			message: 'LSP6: not allowed to run the function.',
+		},
+		'0x3003e7ae': {
+			error: 'NotAllowedERC725YKey(address,bytes32)',
+			message: 'LSP6: not allowed to set the ERC725Y data key.',
+		},
+	},
+	LSP7: {
+		'0xc5e194ab': {
+			error: 'LSP7AmountExceedsBalance(address,address,uint256)',
+			message: "LSP7: token amount exceeds sender's balance.",
+		},
+		'0xf3a6b691': {
+			error: 'LSP7AmountExceedsAuthorizedAmount(address,uint256,address,uint256)',
+			message: "LSP7: token amount exceeds operator's allowance.",
+		},
+		'0x6355e766': {
+			error: 'LSP7CannotUseAddressZeroAsOperator()',
+			message: 'LSP7: cannot authorize address(0) as operator for LSP7 token.',
+		},
+		'0xd2d5ec30': {
+			error: 'LSP7CannotSendWithAddressZero()',
+			message: 'LSP7: cannot send token with address(0).',
+		},
+		'0x263eee8d': {
+			error: 'LSP7InvalidTransferBatch()',
+			message: 'LSP7: invalid transfer batch.',
+		},
+		'0xa608fbb6': {
+			error: 'LSP7NotifyTokenReceiverContractMissingLSP1Interface(address)',
+			message: 'LSP7: token recipient does not implement LSP1 standard.',
+		},
+		'0x26c247f4': {
+			error: 'LSP7NotifyTokenReceiverIsEOA(address)',
+			message: 'LSP7: token recipient is an Externally Owned Account.',
+		},
+	},
+	LSP7CappedSupply: {
+		'0xacf1d8c5': {
+			error: 'LSP7CappedSupplyRequired()',
+			message: 'LSP7CappedSupply: cap supply number required',
+		},
+		'0xeacbf0d1': {
+			error: 'LSP7CappedSupplyCannotMintOverCap()',
+			message: 'LSP7CappedSupply: cannot mint over the max cap supply',
+		},
+	},
+	LSP8: {
+		'0xae8f9a36': {
+			error: 'LSP8NonExistentTokenId(bytes32)',
+			message: 'LSP8: token id does not exist.',
+		},
+		'0x5b271ea2': {
+			error: 'LSP8NotTokenOwner(address,bytes32,address)',
+			message: 'LSP8: tokenOwner given as parameter does not own this token id.',
+		},
+		'0x1294d2a9': {
+			error: 'LSP8NotTokenOperator(bytes32,address)',
+			message: 'LSP8: caller is not an operator for this token id.',
+		},
+		'0x9577b8b3': {
+			error: 'LSP8CannotUseAddressZeroAsOperator()',
+			message: 'LSP8: cannot use address(0) as an operator for token id.',
+		},
+		'0x24ecef4d': {
+			error: 'LSP8CannotSendToAddressZero()',
+			message: 'LSP8: cannot send to address(0).',
+		},
+		'0x34c7b511': {
+			error: 'LSP8TokenIdAlreadyMinted(bytes32)',
+			message: 'LSP8: tokenId already exist (= already minted).',
+		},
+		'0x93a83119': {
+			error: 'LSP8InvalidTransferBatch()',
+			message: 'LSP8: invalid transfer batch.',
+		},
+		'0x4349776d': {
+			error: 'LSP8NotifyTokenReceiverContractMissingLSP1Interface(address)',
+			message: 'LSP8: token recipient does not implement LSP1 standard.',
+		},
+		'0x03173137': {
+			error: 'LSP8NotifyTokenReceiverIsEOA(address)',
+			message: 'LSP8: token recipient is an Externally Owned Account.',
+		},
+	},
+	LSP8CappedSupply: {
+		'0x38d9fc30': {
+			error: 'LSP8CappedSupplyRequired()',
+			message: 'LSP8CappedSupply: cap supply number required',
+		},
+		'0xe8ba2291': {
+			error: 'LSP8CappedSupplyCannotMintOverCap()',
+			message: 'LSP8CappedSupply: cannot mint over the max cap supply',
+		},
+	},
+};
 
 // ----------
 
@@ -234,14 +379,15 @@ const EventSignatures = {
 		/**
 		 * event UniversalReceiver(
 		 *    address indexed from,
+		 * 	  uint256 value,
 		 *    bytes32 indexed typeId,
 		 *    bytes indexed returnedValue,
 		 *    bytes receivedData
 		 * );
 		 *
-		 * signature = keccak256('UniversalReceiver(address,bytes32,bytes,bytes)')
+		 * signature = keccak256('UniversalReceiver(address,uint256,bytes32,bytes,bytes)')
 		 */
-		UniversalReceiver: '0x8187df79ab47ad16102e7bc8760349a115b3ba9869b8cedd78996f930ac9cac3',
+		UniversalReceiver: '0x9c3ba68eb5742b8e3961aea0afc7371a71bf433c8a67a831803b64c064a178c2',
 	},
 	LSP6: {
 		/**
@@ -348,5 +494,7 @@ module.exports = {
 	BasicUPSetup_Schema,
 	ALL_PERMISSIONS,
 	PERMISSIONS,
+	LSP1_TYPE_IDS,
+	Errors,
 	EventSignatures,
 };