+++ contracts/LSP1UniversalReceiver/LSP1UniversalReceiverDelegateUP/Handling/TokenAndVaultHandling.sol
@@ -24,12 +24,12 @@ import {_TYPEID_LSP9_VAULTSENDER} from "../../../LSP9Vault/LSP9Constants.sol";
  * @dev Function logic to add and remove the MapAndArrayKey of incoming assets and vaults
  */
 abstract contract TokenAndVaultHandling {
-    // internal functions
-    function _tokenAndVaultHandling(address sender, bytes32 typeId)
+    function _tokenAndVaultHandling(address caller, bytes32 typeId)
         internal
         returns (bytes memory result)
     {
-        if (sender.code.length == 0) return "";
+        // avoid EOAs spamming the storage
+        if (caller.code.length == 0) return "";
 
         address keyManager = ERC725Y(msg.sender).owner();
         if (!ERC165Checker.supportsERC165Interface(keyManager, _INTERFACEID_LSP6)) return "";
@@ -38,34 +38,43 @@ abstract contract TokenAndVaultHandling {
 
         // check if the caller is the same account controlled by the keyManager
         if (msg.sender != target) return "";
-        (bool senderHook, bytes32 arrayKey, bytes12 mapPrefix, bytes4 interfaceID) = LSP1Utils
-            .getTransferDetails(typeId);
 
-        bytes32 mapKey = LSP2Utils.generateMappingKey(mapPrefix, bytes20(sender));
+        (
+            bool isReceiving,
+            bytes32 arrayLengthKey,
+            bytes12 mapPrefix,
+            bytes4 interfaceID
+        ) = LSP1Utils.getTransferDetails(typeId);
+
+        bytes32 mapKey = LSP2Utils.generateMappingKey(mapPrefix, bytes20(caller));
         bytes memory mapValue = IERC725Y(msg.sender).getData(mapKey);
 
-        if (!senderHook) {
-            // if the map is already set, then do nothing
+        if (isReceiving) {
+            // if the map value is already set, then do nothing
             if (bytes12(mapValue) != bytes12(0)) return "";
+
             result = LSP5Utils.addMapAndArrayKeyViaKeyManager(
                 IERC725Y(msg.sender),
-                arrayKey,
+                arrayLengthKey,
                 mapKey,
-                sender,
+                caller,
                 interfaceID,
                 keyManager
             );
-        } else if (senderHook) {
-            // if there is no map for the asset to remove, then do nothing
+        } else {
+            // if there is no map value for the asset to remove, then do nothing
             if (bytes12(mapValue) == bytes12(0)) return "";
+
+            // if it's a token transfer (LSP7/LSP8)
             if (typeId != _TYPEID_LSP9_VAULTSENDER) {
-                uint256 balance = ILSP7DigitalAsset(sender).balanceOf(msg.sender);
-                // if the amount sent is not the full balance, then do nothing
+                // if the amount sent is not the full balance, then do not remove the keys
+                uint256 balance = ILSP7DigitalAsset(caller).balanceOf(msg.sender);
                 if (balance != 0) return "";
             }
+
             result = LSP5Utils.removeMapAndArrayKeyViaKeyManager(
                 IERC725Y(msg.sender),
-                arrayKey,
+                arrayLengthKey,
                 mapPrefix,
                 mapKey,
                 mapValue,