+++ contracts/LSP1UniversalReceiver/LSP1UniversalReceiverDelegateUP/LSP1UniversalReceiverDelegateUP.sol
@@ -36,6 +36,7 @@ contract LSP1UniversalReceiverDelegateUP is
      */
     function universalReceiverDelegate(
         address sender,
+        uint256 value,
         bytes32 typeId,
         bytes memory data // solhint-disable no-unused-vars
     ) public virtual override returns (bytes memory result) {